backend/requirements.txt
flask
flask-cors
python-dotenv
requests
psycopg2-binary
redis
